+++ src/post-type/query.js
@@ -59,7 +59,7 @@ export const PAGE_QUERY = gql`
 `;
 
 export const PAGE_BY_QUERY = gql`
-  query PageByQuery($uri: String, $pageId: ID) {
+  query PageByQuery($uri: String, $pageId: Int) {
     pageBy(uri: $uri, pageId: $pageId) {
       id
       uri
@@ -72,39 +72,6 @@ export const PAGE_BY_QUERY = gql`
   }
 `;
 
-export const POST_COMMENTS_QUERY = gql`
-  query PostCommentsQuery($id: ID!) {
-    post(id: $id) {
-      id
-      postId
-      title
-      commentStatus
-      comments{
-        nodes {
-          id
-          commentId
-          type
-          content
-          date
-          author{
-            ... on User {
-              id
-              userId
-              nicename
-              avatar {
-                url
-              }
-            }
-            ... on CommentAuthor {
-              id
-              name
-            }
-          }
-        }
-      }
-    }
-  }
-`;
 export const POST_QUERY = gql`
   query PostQuery($id: ID!) {
     post(id: $id) {
@@ -150,7 +117,7 @@ export const POST_QUERY = gql`
 `;
 
 export const POST_BY_QUERY = gql`
-  query PostQuery($postId: ID, $slug: String, $uri: String) {
+  query PostQuery($postId: Int, $slug: String, $uri: String) {
     postBy(postId: $postId, slug: $slug, uri: $uri) {
       id
       postId
@@ -184,102 +151,73 @@ export const POST_BY_QUERY = gql`
       }
       featuredImage {
         id
+        mediaItemId
+        title
+        altText
+        sourceUrl
       }
     }
   }
 `;
 
-/**
- * Mutations
- */
-export const DELETE_COMMENT_MUTATION = gql`
-  mutation DeleteCommentMutation($clientId: String!, $id: ID!) {
-    deleteComment(input: { id: $id, clientMutationId: $clientId }) {
-      clientMutationId
-      comment {
+export const ARCHIVE_QUERY = gql`
+  query ArchiveQuery(
+      $first: Int,
+      $category: String,
+      $tag: String,
+      $year: Int,
+      $month: Int,
+      $day: Int,
+      $author: String,
+      $search: String
+    ) {
+    posts(
+      first: $first,
+      where: {
+        categoryName: $category,
+        tag: $tag,
+        authorName: $author,
+        dateQuery: { year: $year, month: $month, day: $day },
+        search: $search
+      }
+    ) {
+      nodes {
         id
-      }
-    }
-  }
-`;
-
-export const NEW_COMMENT_MUTATION = gql`
-  mutation NewCommentMutation(
-    $author: String,
-    $authorEmail: String,
-    $authorUrl: String,
-    $type: String,
-    $userId: Int,
-    $parent: String,
-    $postId: Int,
-    $content: String!,
-    $date: String!,
-    $clientId: String!,
-  ) {
-    createComment(input: {
-      author: $author,
-      authorEmail: $authorEmail,
-      authorUrl: $authorUrl,
-      type: $type,
-      userId: $userId,
-      parent: $parent,
-      postId: $postId,
-      content: $content,
-      date: $date,
-      clientMutationId: $clientId
-    }) {
-      clientMutationId
-      comment {
-        id,
-        commentId,
-        type,
-        content,
+        postId
+        excerpt
+        content
         date
-        author {
-          ... on CommentAuthor {
+        modified
+        title
+        permalink
+        featuredImage {
+          id
+          mediaItemId
+          title
+          altText
+          sourceUrl
+        }
+        tags {
+          nodes {
             id
             name
-          },
-          ... on User {
-            id
-            nicename
+            slug
           }
         }
-      }
-    }
-  }
-`;
-
-export const UPDATE_COMMENT_MUTATION = gql`
-  mutation UpdateCommentMutation(
-    $type: String,
-    $id: ID!,
-    $content: String!,
-    $date: String!,
-    $clientId: String!,
-  ) {
-    updateComment(input: {
-      type: $type,
-      id: $id,
-      content: $content,
-      date: $date,
-      clientMutationId: $clientId
-    }) {
-      clientMutationId
-      comment {
-        id,
-        commentId,
-        type,
-        content,
-        date
-        author {
-          ... on CommentAuthor {
+        categories {
+          nodes {
             id
             name
-          },
-          ... on User {
-            id
-            nicename
+            slug
+          }
+        }
+        author {
+          id
+          userId
+          nicename
+          avatar {
+            url
+            foundAvatar
           }
         }
       }