+++ src/user/index.js
@@ -4,4 +4,3 @@
 export * from './query';
 export * from './login';
 export * from './user-controls';
-export * from './profile';